Flask-MicroBlog/web/micblog/requirements.txt
Babel==2.3.4
blinker==1.4
click==6.7
coverage==4.3.4
decorator==4.0.11
Flask==0.12
Flask-Babel==0.11.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemyPlus==0.7.5
Flask-WTF==0.14.2
flipflop==1.0
guess-language==0.2
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
packaging==16.8
pbr==1.10.0
pyparsing==2.1.10
python-openid==2.2.5
pytz==2016.10
six==1.10.0
speaklater==1.3
SQLAlchemy==1.1.5
sqlalchemy-migrate==0.10.0
sqlparse==0.2.2
Tempita==0.5.2
Werkzeug==0.11.15
Whoosh==2.7.4
WTForms==2.1
gunicorn==19.6.0
